update notify khi tu choi yeu cau dau gia , accept/deny yeu cau tra hang cho ng ban va nguoi mua
