+++ app/controllers/product.controller.js
@@ -9,8 +9,9 @@ const {userRequestStatus, userWinOrderList, createTitleWinner, createContentWinn
 const {Storage} = require("@google-cloud/storage");
 const {format} = require("util");
 const sse = require("../sse");
-const {updateByWinner} = require("../service/product.service");
+const {updateByWinner, UserReturnProduct} = require("../service/product.service");
 const {da} = require("@faker-js/faker");
+
 
 
 exports.getAuctionHistory = async (req, res) => {
@@ -326,61 +327,26 @@ exports.updateByWinnerController = async (req, res) => {
     }
 }
 
-exports.UserReturnProduct = async (req, res) => {
-    let projectId = process.env.PROJECT_ID // Get this from Google Cloud
-    let keyFilename = 'key.json'
-    const storage = new Storage({
-        projectId,
-        keyFilename,
-    });
-    const bucket = storage.bucket(process.env.BUCKET_NAME); // Get this from Google Cloud -> Storage
-
-    try {
-        const userId = req.userId
-        const productId = req.body?.id
-        if (!req.files || req.files.length === 0) {
-            return res.status(500).send({message: "Please upload at least one file!"});
+exports.UserReturnProductController = async (req, res) => {
+    const result = await UserReturnProduct(req);
+    res.status(result.statusCode).json(result.message);
+    if (!result.error) {
+        const dataForWinner = {
+            title : 'Yêu cầu trả hàng',
+            content : `Yêu cầu trả hàng #${result.data._id.toString()} đã được gửi đến quản trị viên`,
+            url :'',
+            type : 1,
+            receiver : [result.data.winner_id],
         }
-
-        //multifile
-        const uploadPromises = req.files.map(file => {
-            const blob = bucket.file(  Date.now()+ userId + file.originalname);
-            const blobStream = blob.createWriteStream(
-                {resumable: false});
-
-            return new Promise((resolve, reject) => {
-                blobStream.on("error", (err) => {
-                    reject(err);
-                });
-                blobStream.on("finish", async () => {
-                    const publicUrl = format(`https://storage.googleapis.com/${bucket.name}/${blob.name}`);
-                    resolve({url: publicUrl});
-                });
-                blobStream.end(file.buffer);
-            });
-        });
-
-        const results = await Promise.all(uploadPromises);
-        const imageUrls = results.map(item => item.url);
-
-        const returnProduct = await Product.findOneAndUpdate({
-                _id: new mongoose.Types.ObjectId(productId),
-                winner_id: new mongoose.Types.ObjectId(userId),
-                status : 7
-            },
-            {
-                $set: {
-                    status: 9,
-                    'product_delivery.status': 9,
-                    'product_delivery.return_time': new Date(),
-                    'product_delivery.return_image_list': imageUrls,
-                    'product_delivery.return_reason':req.body?.return_reason,
-                }
-            })
-        await returnProduct.save();
-        res.status(200).json(returnProduct)
-    } catch (err) {
-        return res.status(500).json({message: 'DATABASE_ERROR', err})
+        const dataForSeller = {
+            title : 'Yêu cầu trả hàng',
+            content : `Đơn hàng #${result.data._id.toString()} đang được yêu cầu trả lại.`,
+            url :'',
+            type : 1,
+            receiver : [result.data.seller_id],
+        }
+        sse.send( dataForWinner, `returnProductWinner_${result.data.winner_id.toString()}`);
+        sse.send( dataForSeller, `returnProductSeller_${result.data.seller_id.toString()}`);
     }
 }
