+++ app/service/product.service.js
@@ -1,6 +1,9 @@
 const Product = require("../models/product.model");
 const mongoose = require("mongoose");
 const User = require("../models/user.model");
+const {Storage} = require("@google-cloud/storage");
+const {format} = require("util");
+require('dotenv').config()
 
 exports.updateByWinner = async (req, res) => {
     try {
@@ -97,6 +100,75 @@ exports.updateByWinner = async (req, res) => {
             data: [],
             error: true,
             message: "Internal server error.",
+            statusCode: 500,
+        };
+    }
+}
+
+
+exports.UserReturnProduct = async (req) => {
+    let projectId = process.env.PROJECT_ID // Get this from Google Cloud
+    let keyFilename = 'key.json'
+    const storage = new Storage({
+        projectId,
+        keyFilename,
+    });
+    const bucket = storage.bucket(process.env.BUCKET_NAME); // Get this from Google Cloud -> Storage
+
+    try {
+        const userId = req.userId
+        const productId = req.body?.id
+        if (!req.files || req.files.length === 0) {
+            return {
+                data: [],
+                error: true,
+                message: "Please upload at least one file!",
+                statusCode: 500,
+            };
+        }
+
+        //multifile
+        const uploadPromises = req.files.map(file => {
+            const blob = bucket.file(  Date.now()+ userId + file.originalname);
+            const blobStream = blob.createWriteStream(
+                {resumable: false});
+
+            return new Promise((resolve, reject) => {
+                blobStream.on("error", (err) => {
+                    reject(err);
+                });
+                blobStream.on("finish", async () => {
+                    const publicUrl = format(`https://storage.googleapis.com/${bucket.name}/${blob.name}`);
+                    resolve({url: publicUrl});
+                });
+                blobStream.end(file.buffer);
+            });
+        });
+
+        const results = await Promise.all(uploadPromises);
+        const imageUrls = results.map(item => item.url);
+
+        const returnProduct = await Product.findOneAndUpdate({
+                _id: new mongoose.Types.ObjectId(productId),
+                winner_id: new mongoose.Types.ObjectId(userId),
+                status : 7
+            },
+            {
+                $set: {
+                    status: 9,
+                    'product_delivery.status': 9,
+                    'product_delivery.return_time': new Date(),
+                    'product_delivery.return_image_list': imageUrls,
+                    'product_delivery.return_reason':req.body?.return_reason,
+                }
+            })
+        await returnProduct.save();
+        return { data: returnProduct, error: false, message: "Yêu cầu trả hàng được gửi thành công!", statusCode: 200 };
+    } catch (err) {
+        return {
+            data: [],
+            error: true,
+            message: "DATABASE_ERROR!",
             statusCode: 500,
         };
     }
